Delete Employee, Add admin, update
We decided to have admin do maintenance forms, so employee would not be needed. Admin took on that role, and now has the former employee login instead. Admin account creation is done manually in workbench. Made admin_portal.php only accessable to admin. And added a logout button for admins.